fix(PostService.java): rename getPostsByUserId to getPostsOfAuthenticatedUser for better clarity and accuracy fix(PostServiceImplementation.java): remove unused imports and variables fix(PostServiceImplementation.java): remove unnecessary comments and exception handling fix(PostServiceImplementation.java): rename postBelongsToAuthenticatedUser to isPostOwnedByAuthenticatedUser for better clarity feat(PostServiceImplementation.java): add support for checking if post belongs to authenticated user in deletePostById method feat(PostServiceImplementation.java): add support for checking if post belongs to authenticated user in getPostById method feat(PostServiceImplementation.java): add support for checking if post belongs to authenticated user in togglePublicationStatus method feat(PostServiceImplementation.java): add support for getting authenticated user from security context in createPost, getPostsOfAuthenticatedUser, and getAuthenticatedUser methods

+++ blog-backend/src/main/java/com/blog/app/post/service/PostServiceImplementation.java
@@ -1,15 +1,14 @@
 package com.blog.app.post.service;
 
+import com.blog.app.config.security.authentication.JWTAuthentication;
 import com.blog.app.post.dao.PostDao;
+import com.blog.app.post.exceptions.PostNotFoundException;
 import com.blog.app.post.model.Post;
-import com.blog.app.user.model.User;
-import com.blog.app.user.service.UserService;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Service;
 
-import java.security.Principal;
 import java.util.Date;
 import java.util.List;
 import java.util.Objects;
@@ -20,18 +19,16 @@ import java.util.Optional;
 public class PostServiceImplementation implements PostService {
 
     private final PostDao postDao;
-    private final UserService userService;
 
     @Autowired
-    public PostServiceImplementation(PostDao postDao, UserService userService) {
+    public PostServiceImplementation(PostDao postDao) {
         this.postDao = postDao;
-        this.userService = userService;
     }
 
     @Override
     public boolean createPost(Post post) {
-        // todo: set user id from authenticated user
-        post.setUserId(1L);
+        JWTAuthentication authenticatedUser = getAuthenticatedUser();
+        post.setUserId(authenticatedUser.getUserId());
         post.setCreated_at(new Date());
         post.setUpdated_at(new Date());
         // todo: image upload
@@ -45,9 +42,12 @@ public class PostServiceImplementation implements PostService {
 
     @Override
     public boolean deletePostById(Long id) {
-        // todo: check if post belongs to the authenticated user
-        if (!postBelongsToAuthenticatedUser(id)) {
-            // todo: throw exception
+        Optional<Post> optionalPost = postDao.getPostById(id);
+        if (optionalPost.isEmpty()) {
+            throw new PostNotFoundException("Post not found");
+        }
+        if (!isPostOwnedByAuthenticatedUser(optionalPost.get())) {
+            throw new PostNotFoundException("Post not found");
         }
         return postDao.deletePostById(id);
     }
@@ -58,24 +58,50 @@ public class PostServiceImplementation implements PostService {
     }
 
     @Override
-    public List<Post> getPostsByUserId(Long userId) {
-        // todo: return post of authenticated user
-        return null;
+    public List<Post> getPostsOfAuthenticatedUser(Long userId) {
+        JWTAuthentication authenticatedUser = getAuthenticatedUser();
+        postDao.getPostsByUserId(authenticatedUser.getUserId());
+        return postDao.getPostsByUserId(authenticatedUser.getUserId());
     }
 
     @Override
     public List<Post> getPostsByUsername(String username) {
-        return postDao.getPostsByUsername(username);
+        JWTAuthentication authenticatedUser = getAuthenticatedUser();
+
+        if (authenticatedUser.getUsername().equals(username)) {
+            return postDao.getAllPostsByUsername(username);
+        }
+
+        return postDao.getPublicPostsByUsername(username);
     }
 
     @Override
     public Optional<Post> getPostById(Long postId) {
-        return postDao.getPostById(postId);
+        Optional<Post> optionalPost = postDao.getPostById(postId);
+        if (optionalPost.isEmpty()) {
+            throw new PostNotFoundException("Post not found");
+        }
+        Post post = optionalPost.get();
+
+        if (!post.isPublish() && !isPostOwnedByAuthenticatedUser(post)) {
+            throw new PostNotFoundException("Post not found");
+        }
+
+        return Optional.of(post);
     }
 
     @Override
     public boolean togglePublicationStatus(Long postId) {
-        // todo: check if post belongs to the authenticated user
+        Optional<Post> optionalPost = this.postDao.getPostById(postId);
+
+        if (optionalPost.isEmpty()) {
+            throw new PostNotFoundException("Post not found");
+        }
+
+        if (!isPostOwnedByAuthenticatedUser(optionalPost.get())) {
+            throw new PostNotFoundException("Post not found");
+        }
+
         return false;
     }
 
@@ -85,32 +111,33 @@ public class PostServiceImplementation implements PostService {
      * @param post The post to be checked.
      * @return true if the post belongs to the authenticated user, false otherwise.
      */
-    private boolean postBelongsToAuthenticatedUser(Post post) {
+    private boolean isPostOwnedByAuthenticatedUser(Post post) {
         try {
-            // todo : throw exception if user is not authenticated
             log.info("Checking if post belongs to authenticated user");
-            Principal principal = SecurityContextHolder
-                    .getContext().getAuthentication();
-            String email = principal.getName();
-            Optional<User> optionalUser = userService.findUserByEmail(email);
-
-            if (optionalUser.isEmpty()) {
-                return false;
-            }
-
-            User user = optionalUser.get();
-
-            return Objects.equals(user.getId(), post.getUserId());
+            JWTAuthentication auth = getAuthenticatedUser();
+            return Objects.equals(auth.getUserId(), post.getUserId());
         } catch (Exception e) {
             log.error("Error checking if post belongs to authenticated user");
             return false;
         }
     }
 
-    private boolean postBelongsToAuthenticatedUser(Long postId) {
-        Optional<Post> optionalPost = getPostById(postId);
-        if (optionalPost.isEmpty()) return false;
-        return postBelongsToAuthenticatedUser(optionalPost.get());
+    /**
+     * Obtains the authenticated user from the security context.
+     *
+     * @return The authenticated user's JWTAuthentication object.
+     * @throws RuntimeException if the user is not authenticated.
+     */
+    private JWTAuthentication getAuthenticatedUser() {
+        log.info("Getting authenticated user");
+        JWTAuthentication principal = (JWTAuthentication) SecurityContextHolder
+                .getContext().getAuthentication();
+
+        if (principal == null) {
+            throw new RuntimeException("User not authenticated");
+        }
+
+        return principal;
     }
 
 }
